requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
groq
httpx==0.27.2
pillow==10.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.32.3
